Check tiledb query status in read_current_batch
This change alters the behavor of read_current_batch to properly handle
cases we the tiledb vcf query is incomplete but we've moved to a new
underlying tiledb core query. Previously we incorrectly checked the
query_results which showed a complete query, ignoring that we had
allocated and setup a new Query object and that the query_results was
invalid.