Reset current buffer sizes only in read
Resetting these in read_current_batch causes results to
be overridden if we hit a case where the tiledb query
is completed but the vcf read is not AND the user's
buffers are not filled. In that cases we previously
would (correctly) start a new tiledb query but incorrectly
reset the buffers, thus overriding the last bit of
results from the previously completed tiledb query.