Add support for enabling and dumping TileDB Stats
Support is added to the CLI, Python, Java, Spark and the C-API

The CLI will print statistics at the end of an export if enabled. Python
has a new function on the TileDBDataSet class for returning statistics
as a string. Spark will print the stats to the log level specified at
the end of a partition read. The C-API also has the option for
returning the statistics as a char**.