Fix over-allocating attribute buffers
When VCF is given a memory budget of N, N/2 is given to the core while
the other N/2 are given to the VCF reader.

The VCF reader allocates M attribute buffers, where each buffer should have
a size of (N/2)/M, but it currently is allocated N/2 for _each_ buffer,
which over-allocates the VCF reader by a scale of M.

This fixes the issue.