VCF record iteration for overlapping records
Currently, we buffer VCF records into memory and store the minimum start
position and maximum end position of these buffered records. When we iterate
to buffer the next set of VCF records, we seek to the maximum end position
and read from there. This breaks if there are any overlapping records.

This patch chanes the behavior from open+seek on each iteration to opening
once on the first iteration and making use of the internal htslib iterator
to move between records.