Add htslib plugin for reading files via VFS
The new plugin loads files that use the schema of `vfs://`. Normal
URIs such as `s3://my_bucket/regions.bed` become
`vfs://s3://my_bucket/regions.bed`. This is handled internally in
libtiledbvcf. As a use you pass in the normal URI still.

This plugin is written in C and designed to be upstreamed. Helper
functions are written in c++ to initialize the plugin and to set the
tiledb config/context from the reader/writer or cli. This ensures that
any TileDB-VCF parameters set for TileDB config are used by htslib in
reading files.

The plugin itself is not thread-safe but the c++ utility functions, and
we only interact with the plugin initializing through the c++ functions.