Heap-node ownership of buffered records
This is a prerequisite to ingesting overlapping records.

Currently, the `VCF` class iterates through a VCF file with a readahead that
buffers multiple records into a single buffer. When the last record in the
buffer is read, the buffer is cleared and populated with the next set of
records. This is OK because we currently enforce that all records are ordered
and without overlap.

To support overlapping records, records may be processed out-of-order. We can
not safely clear the single buffer because nodes on the heap may still point
to them. This change moves ownership of the records from the `VCF` instance
to the individual nodes on the heap.

The `VCF` still performs a readahead, but places each record onto a queue. The
writer worker will pop the next record and place it into the heap. When the
queue is empty, the `VCF` instance will perform the next readahead. There is
no additional copying and the read IO is unchanged. As an optimization,
when the end node on a heap is popped, the allocated record is returned to the
`VCF` instance for re-use.

This patch is verbose because the `VCF` instance had to be split into a V2
and V3 variant to support backwards compatibility. Shared static utilities
have been placed into a `vcf/vcf_utils.h` file.