Fix segfault in InMemoryExporter::get_info_fmt_value
We encode and store INFO values in a buffer as: keystr,type,nvalues,values

This fixes an off-by-one where we read an extra INFO value from the query
result. If the overread memory is zeroed, the behavior works as expected. If
the memory is not-zeroed, we end up with a segfault. I could also envision
corrupt/false values being read, but I have not explicitly observed that
scenario.