Error in spark for sample option and filter
If a user passes both a sample option (`sampleFile` or `samples`) and
then they try to add a predicated filter in a where clause return an
error to the user. When both a sample option and a predicate are passed
we can't know which (or all) of the samples the user was trying to
access. The safest option is to error and force the user to set either a
where clause or an option only.