Delay creating contexts for read
There is no need to call init_tiledb() and recreate the context for all
possibly parameter change functions. These calls will be done before the
dataset is opened and a reader returned. Avoiding these calls avoids a
problem of initializing TBB twice with different thread counts.

init_tiledb() is left for the bed_file and sample_file setters as those
need VFS enabled to validate the files exist. These functions should
only be called after the dataset is opened and a reader object is
returned, so the code paths should never hit in normal api usage.