Add support for reading BED file in parallel.
If a bed file is passed and it has an tabix index we will parallelize
the read and parsing based on the chromosomes in the index.

This also change the htslib plugin to mark files as remote. Since we
only use this for the BED file currently that is okay. The problem is
when you mark files as local htslib tries to stat the index file which
fails.