Limit region comparison to contig
A new unordered map is used to keep an index record of all regions for a
contig. This introduces one hash comprison for the map lookup, however a
string comparison was already required for contig matching. As a result
the reduced search domain likely outweighs the performance hit of a hash
vs the string comparison.