Every record needs to independently check regions
We can not move the minimum region index as we iterate TileDB query
results because the results are unordered. Instead we need to check each
cell against all relevant regions. Relevant regions are determined by
doing a binary search to find the first possible region that will
intersect. The algorithm is unchanged after the binary search.

This change effects both v2 and v3 read algorithms