Update v2/v3 reads to use new intersection algo
There is a bug in the existing intersection reads, in that we don't
properly limit exporting duplicate records from anchors. The
last_exported_end does not correctly handle anchors, and has problems
because TileDB query results are unordered. Instead a new formula for
checking intersection which is more generic and guaranteed to not have
issue with anchors or duplicates.