Switch to using point ranges for all sample ids
Instead of issueing a range query for all samples between the min id and
max id, we instead will issue points for all samples. In TileDB 2.1
point ranges can be coalesced automatically. In the worst case before if
a user issues a query for 2 samples, the first and last TileDB-VCF would
being in the entire dataset. Now it will just be the two samples needed.