v2/v3 arrays contig check needs to lower side also
Since TileDB records are unordered we need to check not only if a new
contig has been reached on the upper bound but also lower bound.
Previous we only checked if the end moved to the next contig. However
its possible for a record to be on a previous contig since the records are
unordered.