Validate region has same contig as record
For v2/v3 arrays we need to validate region is same contig as record by
checking the names. Even with global ordering in v2/v3 arrays its
possible for a region to incorrectly overlap because a record extends
before or after a contigs boundaries. Contig offsets are not guaranteed
to be unique so we use a string comparison of the contig name.