Optimize v2/v3 read queries
This solves a performance regression introduced in v0.6 when the read
algorithm was improved. The changes are too move the string comparision
of contig until after we've checked all integer based conditions.
Second, and more impactful, exit the intersection check early if the
regions have moved pasted the record. The regions are sorted so this
early exit is safe.