Remove contig check from process_query_results_v4
Since TileDB queries now run on a single contig, we no longer need to
check each record to make sure it matches the previous records contig,
and optionally lookup the new contig regions. Instead we can lookup
contig regions once before we process the cells. We can now also remove
the need to build the contig string in the looping of the cells.