Build Chunked Arrow Table in _read_partition()
The Dataset read() and continue_read() functions take in a return_type
parameter that allows specifying whether the results are contained in a Pandas
dataframe or Arrow table. The Dataset continue_read() function also takes in
a release_buffers parameter that indicates whether the previous results should
be freed before continuing the read.