Fix native library building for tags
This drops the `succeeded()` condition on the `BuildNativeLibs` stage.
The `succeeded()` requires all previous stages to succeed which does not
work for tags because the first CI stage of the unit tests are skipped.
This means we are not building the multi-lib jar for spark and instead
are only building linux. This commit solves this problem.