Add threshold for double buffering of TileDB Query
This add an internal threshold for double buffering of the TileDB query.
If the buffers are too small, then the overhead of splitting further and
running two queries is large. As a heuristic we use a size of 200MB (that
would be 100MB per buffer in double buffering) as the threshold. If the
buffers are smaller than this we disable double buffering and only run a
single query.