Automatically set the TileDB memory budget config
This changes the TileDB memory budget parameters to be set on a
configuration option that is now attached to the query. TileDB 2.2.2
introduced the ability to set a config object on the query which allows
certain parameters to be set more localized. This includes the
`sm.memory_budget` and `sm.memory_budget_var` parameters. We now will
set this to the same size as the query buffer sizes.