Split TileDB query buffers and Query Memory Budget
TileDB might use up to the `sm.memory_budget` and `sm.memory_budget_var`
bytes for internal memory when reading the tiles into memory before
decompression into the user buffers. This means for peak potential
memory usage we need to account for the internal tiledb memory in the
buffer sizes.