Tile cache is fixed percentage of memory_budget_mb
This change makes it so that we compute the size of the tile cache as
10% of the VCF memory budget. This will limit the size of the memory
usage as a user would expect. Previously the tile cache was set to 1GB
or user configured, which would be mean if the user hadn't set it they
might end up using more than expected memory.