Give TileDB Core 75% of buffer budget
In most cases TileDB core library will need to decompress more
tiles/chunks to filter to the data requested then will be set in the
query buffers. This means than an even split is likely to cause more
incomplete queries. As such, switch from a 50/50 split to a 75/25 split
will yield performance improvements.