Let htslib manage info/fmt field count on export
For exports to vcf/vcf we should let htslib manage the info/fmt field
counts. As we add the fields on export, htlib increments the account
automatically. Manually override it is typically fine, except we were
adding an extra count for the END field which might not always exist.
The proper solution is to just stop override the field counts.