Preemptively clear buffers when read is complete
This allows for the query buffers in C++ to free their memory when the
user might hang onto the reader class. This can happen in python if the
user keeps the Dataset class in the current scope. In spark there can
also be a delay due to GC in cleaning up the reader. Once the
TileDB-VCF query is complete it is safe to free the TileDB query
buffers.