Always reset the TileDB Query buffers
When we are not double buffering we need to reset the query buffers so
that the 'buffer size' as reported by the core library is the real size.
This fixes an edge case when the core library is serializing the query
that could cause the query buffer size to shrink each time a request was
made. The problem was that the core library would send the buffer size
instead of original buffer size to TileDB Cloud.

This works around the core issue but always setting the buffers. The
core bug will be fixed in a pending release.