Lazily fetch all sample names for v4 arrays
On dataset open we not longer fetch all sample names from the VCF header
array. Instead of only fetch if the user request the list of samples.
The check for if a sample exists in the dataset when a query is issued
is now optional and disabled by default. This removes the need to
preload the entire sample list for v4 arrays.