Adjust info/fmt field mapping to be lazily loaded
We will now attempt to only load the info/fmt field map when it is first
needed. We also will use an existing sample header from the query
instead of fetching the first one from the header array. This has a
bonus effect of now allowing different headers to co-exist and the
relevant one to the query can be used.

If the user queries the queryable attributes, or materialized attributes
this will force a building of the info/fmt field mapping. This typically
only effects spark where we have to get the full schema to provide to
the spark API before the user runs a query. Users in python also can
fetch the list of queryable attribute but this is usually done async for
data exploration.