Use fixed buffer budget for VCF headers
This removes the use of est_result_size for VCF header queries. There is
no need to use the estimate because it can have higher overhead then the
query here. The VCF headers are usuablly small and estimating the size
and ignoring the user set memory budget is not worth the trade off.
Instead it is better to honor the user's memory choices and fetch with
fixed sized buffers.