Fix attribute_is_fixed_len for v4 sample
The `attribute_is_fixed_len` function previously was returning fixed
length for `sample` even for v4 arrays because the name is the same
between v2/v3 and v4. This adjusts this to check for v2/v3 cases.

This only effected the buffer size building, and caused us to be off by
missing 1 set of offsets in the calculations for max sizes we could use.