Fix partitioning logic for uneven division
There was a bug in the logic of partitioning in that we used ceil not
floor when computing the number of element per partition. This meant
that if we rounded up and used up all the elements before the requested
user partitioning count we would run out of samples/regions and end up
with empty partitions. The empty partitions were not handled gracefully
and caused the library to incorrectly fetch all headers. Depending on
the exact scenario it might have also resulted in a segfault when
attempting to set empty ranges on the core library due to accessing
non-existent vector elements.

The fix for this issue is to use floor when computing element per
partition and to just assign the left overs to the first N partitions,
where one is given to each partition until we run out of "left over"
elements.