Fix concurrent issues with dataset destructor
This introduces a new rwlock and thread TileDBVCFDataset attribute to
handle race conditions when destroying the dataset class. If the
preloading of the non empty domains had not completed by the time the
destructor was called this could cause a segfault.

Similarly we now also grab read locks of the data array or vcf header
array rwlock in order to block the destructor if another thread was
making use of the array objects. No where in the c++ code do we
currently access in parallel to destruction, however in order to be
thread-safe in the c-api this is needed.