Support for merging contigs into single fragments
This is a (large) performance optimization in cases where the samples
have a large number of pseudo configs which are small. These small
contigs create a large number of fragments which can cause slower
listing of the array on S3/GCS/Azure Blob Store object stores. This
makes opening the array take significantly longer in certain scenarios.
The change here allows the user to set a whitelist/blacklist of contigs
which are allowed to be merged into a single fragment in the sample
batch.

The default blacklist will keep human contigs in USCS or ensembl format
from being merged. For performance of reads, it is desirable to keep
contigs which are going to read frequently or are large in size separate
and not merged.

Further merge takes into account (automatically) lexicographical
ordering. That is in the event of chromsomes `chr1`, `chr1_x`, `chr1_y`,
`chr2`, and `chr2_x`, we will merge `chr1_x` and `chr1_y` together but
not merge `chr2_x` with `chr1_y` because we must avoid overlapping the
fragment's non empty domain with `chr2`.