Update TileDB Embedded to 2.28.1 (#826) (#827)
* Bump nightly build release branch to 2.28 (#819)

* fix img and academy links (#815)

Use TileDB Academy links for docs

* remove ENTRYPOINT for Nextflow compatibility (#816)

* remove ENTRYPOINT for Nextflow compatibility

* add tiledbvcf to test_cmd since no entrypoint

* Updated spdlog to version 1.15.3 (#821)

This required adding an include of fmt/ranges.h because of a bug in libfmt 11, which spdlog depends on.

* Revert "Updated spdlog to version 1.15.3 (#821)" (#824)

This reverts commit e4b00ac726a271d2b53e246db79f2fee0d93613f.

* Updated HTSlib to version 1.22.1 (#822)

* Updated HTSlib to version 1.22.1

This includes changing the CMake External Project URL hash for HTSlib to use a SHA256 checksum because GitHub now auto-generates SHA256 checksums for HTSlib release artifacts.

* Removed unused HTSlib CMake patch

* Updated Azure CI config to install curl prior to bcftools

bcftools 1.22 contains a bug fix that causes the build in this configuration to fail if curl is not present.

* Update to TileDB Embedded 2.28.1

---------

Co-authored-by: John Blischak <jdblischak@gmail.com>
Co-authored-by: Jeremy Leipzig <jeremy.leipzig@tiledb.com>