Deletion memory usage (#835)
* Sample stats are now updated for one sample at a time during deletion

Previously there was a bug where the stats for every sample being deleted were processed repeatedly as each sample was deleted.

* Export parameters can now be passed to the dataset class's delete method

This allows the reader used for this operation to be configured from outside the method, e.g. via the CLI. The previous method signature is preserverd as a wrapper for the new signature.

* The deletion memory budget can now be controlled via the CLI

* Applied clang-format

* Added CLI deletion tests

* Added multi sample deletion unit test