Bring 0.38 up to date with main (#836)
* Bump nightly build release branch to 2.28 (#819)

* fix img and academy links (#815)

Use TileDB Academy links for docs

* remove ENTRYPOINT for Nextflow compatibility (#816)

* remove ENTRYPOINT for Nextflow compatibility

* add tiledbvcf to test_cmd since no entrypoint

* Updated spdlog to version 1.15.3 (#821)

This required adding an include of fmt/ranges.h because of a bug in libfmt 11, which spdlog depends on.

* Revert "Updated spdlog to version 1.15.3 (#821)" (#824)

This reverts commit e4b00ac726a271d2b53e246db79f2fee0d93613f.

* Updated HTSlib to version 1.22.1 (#822)

* Updated HTSlib to version 1.22.1

This includes changing the CMake External Project URL hash for HTSlib to use a SHA256 checksum because GitHub now auto-generates SHA256 checksums for HTSlib release artifacts.

* Removed unused HTSlib CMake patch

* Updated Azure CI config to install curl prior to bcftools

bcftools 1.22 contains a bug fix that causes the build in this configuration to fail if curl is not present.

* Update TileDB Embedded to 2.28.1 (#826) (#827)

* Bump nightly build release branch to 2.28 (#819)

* fix img and academy links (#815)

Use TileDB Academy links for docs

* remove ENTRYPOINT for Nextflow compatibility (#816)

* remove ENTRYPOINT for Nextflow compatibility

* add tiledbvcf to test_cmd since no entrypoint

* Updated spdlog to version 1.15.3 (#821)

This required adding an include of fmt/ranges.h because of a bug in libfmt 11, which spdlog depends on.

* Revert "Updated spdlog to version 1.15.3 (#821)" (#824)

This reverts commit e4b00ac726a271d2b53e246db79f2fee0d93613f.

* Updated HTSlib to version 1.22.1 (#822)

* Updated HTSlib to version 1.22.1

This includes changing the CMake External Project URL hash for HTSlib to use a SHA256 checksum because GitHub now auto-generates SHA256 checksums for HTSlib release artifacts.

* Removed unused HTSlib CMake patch

* Updated Azure CI config to install curl prior to bcftools

bcftools 1.22 contains a bug fix that causes the build in this configuration to fail if curl is not present.

* Update to TileDB Embedded 2.28.1

---------

Co-authored-by: John Blischak <jdblischak@gmail.com>
Co-authored-by: Jeremy Leipzig <jeremy.leipzig@tiledb.com>

* Improve run-time performance of delete exporter (#833)

* Added optional field filters to recover_record(...) method in Exporter class

This allows what (materialized) info and format fields are added to the BCF record to be limited to a specific set.

* Updated export_record(...) method in DeleteExporter class to only export genotype fields

This is to improve run-time performance.

* Update the Buffer class data() method to compute string view directly from offsets (#834)

Co-authored-by: Julia Dark <julia@tiledb.com>

* Deletion memory usage (#835)

* Sample stats are now updated for one sample at a time during deletion

Previously there was a bug where the stats for every sample being deleted were processed repeatedly as each sample was deleted.

* Export parameters can now be passed to the dataset class's delete method

This allows the reader used for this operation to be configured from outside the method, e.g. via the CLI. The previous method signature is preserverd as a wrapper for the new signature.

* The deletion memory budget can now be controlled via the CLI

* Applied clang-format

* Added CLI deletion tests

* Added multi sample deletion unit test

---------

Co-authored-by: John Blischak <jdblischak@gmail.com>
Co-authored-by: Jeremy Leipzig <jeremy.leipzig@tiledb.com>
Co-authored-by: Julia Dark <julia@tiledb.com>