Added a scan_all_samples parameter to read_variant_stats() and read_variant_stats_arrow()
This allows variant stats to be computed relative to all samples in the dataset. To support this change, an AFComputer abstract class and two subclasses were introduced to encapsulate the AF stats code. Additionally, supporting types were introduced and variables that use these types were renamed to clarify how the code that uses these classes works.