Added a "regions" parameter to read_allele_count() and read_allele_count_arrow()
This allows multiple regions to be queried simultaneously. Consequently, a "contig" column was added to the returned DataFrame and Table, respectively, to clarify what contig positions occur on. The old "region" parameter is still present but has been deprecated in preference for the new "regions" parameter. Also, the code for parsing and preparing input regions used by read_variant_stats() and read_variant_stats_arrow() was encapsulated in a _prepare_regions() method so that it may be used in this context as well.