Add pre-commit (#853)
* Added a pre-commit configuration

The configuration runs the formatters currently used by the project, i.e. clang-format v15 and black v25.

* Removed the format commands from the Makefile

These will now be handled automatically with pre-commit and GitHub workflows.

* Added a GitHub workflow that runs pre-commit

* Removed Clang Tools from SuperBuild

This module was already deprecated and is being removed since ClangFormat is now managed by pre-commit.

* Removed formatting from existing CI

This is because formatting is now handled via the pre-commit CI. Note that pylint was removed here as well since this is the only place in the project that it is used.

* Removed run-clang-format.sh CI scripts

This has been replaced by pre-commit.

* Updated clang-format pre-commit configuration to include Java API

* Removed clang-format from Java JNI script

* Removed black from Python API dependencies

This is because black is now managed by pre-commit.

* Update formatting related documentation

Specifically, clang-format was dropped from the Java API documentation and pre-commit in now mentioned in the contributing documentation.

* Corrected external sources directory path in pre-commit configuration