Improve API for VCF read_variant_stats() and read_allele_count() (#852)
* Updated Dataset.read_variant_stats() to return a pandas.DataFrame

Previously the method returned a pyarrow.Table but the return type hint was a pandas.DataFrame, so the new behavior is correct. A pyarrow.Table can still be retrieved via a new Dataset.read_variant_stats_arrow() method.

* Added a "regions" parameter to read_variant_stats() and read_variant_stats_arrow()

This allows multiple regions to be queried simultaneously. Consequently, a "contig" column was added to the returned DataFrame and Table, respectively, to clarify what contig positions occur on. The old "region" parameter is still present but has been deprecated in preference for the new "regions" parameter.

* Updated read_variant_stats() and read_variant_stats_arrow() to sort and consolidate regions

This ensures that stats have a consistent ordering and that there isn't redundant rows when regions overlap. Note that a Region class was introduced to facilitate this functionality.

* Added a scan_all_samples parameter to read_variant_stats() and read_variant_stats_arrow()

This allows variant stats to be computed relative to all samples in the dataset. To support this change, an AFComputer abstract class and two subclasses were introduced to encapsulate the AF stats code. Additionally, supporting types were introduced and variables that use these types were renamed to clarify how the code that uses these classes works.

* Added a drop_ref parameter to read_variant_stats() and read_variant_stats_arrow()

This causes the "ref" allele to be omitted from the results.

* Updated Dataset.read_allele_count() to return a pandas.DataFrame

Previously the method returned a pyarrow.Table but the return type hint was a pandas.DataFrame, so the new behavior is correct. A pyarrow.Table can still be retrieved via a new Dataset.read_allele_count_arrow() method.

* Added a "regions" parameter to read_allele_count() and read_allele_count_arrow()

This allows multiple regions to be queried simultaneously. Consequently, a "contig" column was added to the returned DataFrame and Table, respectively, to clarify what contig positions occur on. The old "region" parameter is still present but has been deprecated in preference for the new "regions" parameter. Also, the code for parsing and preparing input regions used by read_variant_stats() and read_variant_stats_arrow() was encapsulated in a _prepare_regions() method so that it may be used in this context as well.

* Made AlleleCountReader a friend of AlleleCount

This gives AlleleCountReader access to AlleleCount's private members, thus removing the need for redundant variables.

* Updated AlleleCountReader to use AlleleCount's private ALLELE_COUNT_ARRAY member during construction

* Updated other Dataset methods to use new Region functionality

This means all regions passed to the API are now handled consistently, i.e. they are parsed, sorted, and consolidated the same way for every method.

* Explicitly sorting results of read_variant_stats() and read_allele_count()

This is to provide a more deterministic ordering than what is provided by the underlying C++ API.

* Updated all pyarrow.Table.to_pandas() calls to reduce memory consumption

This was done using the split_blocks and self_destruct parameters in tandem.